Fix clientapp config files for Jest, Karma and Cypress.
Overview of changes:
A new clientapp had been created using 'dotnet new angular' when
the solution was re-organized to follow the 'Clean Architecture'
pattern. In doing so, we lost the configuration that existed in
the prior clientapp. This configuration allowed for unit testing using
either karma or Jest and also for integration testing with cypress.