Remove videogular & amcharts. Docker build issue.
1. working on failure of clientapp docker build.
It was failing with a "killed" message during stage 1.
I removed stage 2 and did a "CMD sh" after the
copy of source files in stage 1
so I could enter the container at this spot and
do the build manually. The problem is memory maxes out.

2. Remove videogular. Have Staging mimic production.
Videogular had been replace by video.js. The code
hadn't been removed til now in case of problems with video.js.
Changed augular.json so that theStaging configuration mimics
Production. This is as it should be. Plus, the fact that Staging
was producing sourceMaps is one of the reasons for the memory
issue in the current build.

3. Add webpack-bundle-analyzer to clientapp
   for debugging the memory maxing out on "ng build:staging"

4. Removed amcharts